fix a case for switching back to char screen; adjusted EventHandler with a BlockingQueue; improve Shop packets and handling; added auto pick task

+++ game-server/src/main/java/com/jftse/emulator/server/core/packets/matchplay/S2CMatchplayEndBasicGame.java
@@ -4,7 +4,7 @@ import com.jftse.server.core.protocol.PacketOperations;
 import com.jftse.server.core.protocol.Packet;
 
 public class S2CMatchplayEndBasicGame extends Packet {
-    public S2CMatchplayEndBasicGame(short resultTitle) {
+    public S2CMatchplayEndBasicGame(byte resultTitle) {
         super(PacketOperations.S2CMatchplayEndBasicGame);
 
         this.write(resultTitle); // 0 = Loser, 1 = Winner
